cmake: Use correct OpenGL library handle.
To link OpenGL it is needed to use `OpenGL::GL` instead of
`OpenGL::OpenGL`, as can be seen in `FindOpenGL.cmake`.

Otherwise the following error is generated on CMake 3.13.3:
```
  Target "ktx2any" links to target "OpenGL::OpenGL" but the target was not
  found.  Perhaps a find_package() call is missing for an IMPORTED target, or
  an ALIAS target is missing?
This warning is for project developers.  Use -Wno-dev to suppress it.
```

This also leads to linker errors later. This patch fixes this.